requirements.txt
ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
gradio>=4.20.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0
albumentations>=1.3.0
